requirements.txt
Django>=5.0
djangorestframework>=3.15
drf-spectacular>=0.27
django-prometheus>=2.3
django-filter>=24.3
gunicorn>=21.2
whitenoise>=6.6
pytest-django>=4.8
